modal cleanup
