+++ frontend-vite/src/components/CreateIssueModal.jsx
@@ -1,68 +1,214 @@
 import { useDispatch, useSelector } from 'react-redux'
-import { addTitle, addDescription } from "../store/issueReducer"
 import TypeDropdown from './TypeDropdown.jsx'
 import PriorityDropdown from './PriorityDropdown.jsx'
+// import axiosInstance from '../util/AxiosInstance.js'
+import axios from 'axios'
+import { Fragment, useRef, useState } from 'react'
+import { Dialog, Transition } from '@headlessui/react'
+import { ExclamationTriangleIcon } from '@heroicons/react/24/outline'
+
 
 const CreateIssueModal = ({closeModal}) => {
-    const dispatch = useDispatch()
+  const [title, setTitle ] = useState()
+  const [description, setDescription ] = useState()
+  const [comment, setComment ] = useState()
+  const [priority, setPriority] = useState()
+  const [type, setType ] = useState()
 
-    const handleTitle = (e) => {
-        e.preventDefault()
-        dispatch(addTitle(e.target.value))
-    }
+  // const currentUser = useSelector(state => state.user.currentUser)
 
-    const handleDescription = (e) => { 
+  const testUser = {
+      "id": "641ba87494ba927d1a1e932c",
+      "email": "test1@gmail.com",
+      "firstName": "Test",
+      "lastName": "One",
+      "role": "ROLE_USER",
+      "avatarUrl": "https://chinguboard-dev.s3.us-east-2.amazonaws.com/f805ce7c-5dba-46f3-be54-90c27983aacc_29348169ecc5b8d01ac28beb2c5a4a79.png"
+  }
+
+  const testIssue = {
+    "title": "testTitle",
+    "description": "testDescription",
+    "assignees": [],
+    "comments": [],
+    "createdBy": testUser,
+    issueType: "TASK",
+    priority: "LOW",
+    status: "BACKLOG",
+  }
+  
+  const handleTitle = (e) => {
         e.preventDefault()
-        dispatch(addDescription(e.target.value))
-    }
-    
-    const handleSave = (e) => {
+        setTitle(e.target.value)
+  }
+  
+  const handleDescription = (e) => { 
         e.preventDefault()
-        // make POST request and close modal
+        setDescription(e.target.value)
+  }
+
+  const handleComment = (e) => { 
+        e.preventDefault()
+        setComment(e.target.value)
+  }
+
+  const handlePriority = (priority) => {
+    setPriority(priority)
+  }
+
+  const handleType = (type) => {
+      setType(type)
+  };
+
+  const handleSave = (e) => {
+        e.preventDefault()
+        // console.log(title, description, comment, priority, type, userEmail)
+        postIssue(testIssue)
         closeModal(false)
+  }
+  
+  const postIssue = async (testIssue) => {
+      try {
+        const response = await axios.post(`http://localhost:8080/api/issues/create?projectId=641ba8e494ba927d1a1e932d`, testIssue);
+        return response.data;
+      } catch (error) {
+        console.log("error", error)
+      }
     }
 
     return (
-        <div className="modalBackground">
-            <div className="modalContainer">
-                <button onClick={() => closeModal(false)}>X</button>
-                <div className="title">
-                    <h1>Create Issue</h1>     
-                </div>
-                <div className="body"> 
-                    <form>
-                        <label>
-                        Title:
-                            <input type="text" name="name" onChange={handleTitle}/>
-                        </label>
-                        <label>
-                        Description:
-                            <input type="text" name="name" onChange={handleDescription}/>
-                        </label>
-                        <label>
-                        Assignees:
-                            <input type="text" name="name" onChange={handleDescription}/>
-                        </label>
-                        <label>
-                        Comments:
-                            <input type="text" name="name" onChange={handleDescription}/>
-                        </label>
-                        <div>
-                            <TypeDropdown />
-                        </div>
-                        <div>
-                            <PriorityDropdown />
-                        </div>
-                    </form>
-                    <p>Created by: {useSelector(state => state.user.email)} </p>
-                </div>
-                <div className="footer">
-                    <button onClick={handleSave}> Save </button>    
-                 </div>
+    <div className="fixed z-10 inset-0 overflow-y-auto">
+    <div className="flex items-end justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0">
+      <div className="fixed inset-0 transition-opacity" aria-hidden="true">
+        <div className="absolute inset-0 bg-gray-500 opacity-75"></div>
+      </div>
+      <span className="hidden sm:inline-block sm:align-middle sm:h-screen" aria-hidden="true">&#8203;</span>
+      <div className="inline-block align-bottom bg-white rounded-lg text-left overflow-hidden shadow-xl transform transition-all sm:my-8 sm:align-middle sm:max-w-lg sm:w-full">
+             <form className="mt-6">
+           <div className="mb-2">
+             <label>
+              <span className="text-gray-700">{useSelector(state => state.user.email)}</span>
+               <input
+                type="text"
+                name="name"
+                className="
+
+            w-full
+            block px-16 py-2 mt-2
+            border-gray-300
+            rounded-md
+            shadow-sm
+            focus:border-indigo-300
+            focus:ring
+            focus:ring-indigo-200
+            focus:ring-opacity-50
+          "
+                placeholder="Title"
+                onChange={handleTitle}
+              />
+            </label>
+          </div>
+          <div>
+                <TypeDropdown handleType={handleType} />
             </div>
+            <div>
+                <PriorityDropdown handlePriority={handlePriority}/>
+            </div>
+          <div className="mb-2">
+            <label>
+              <span class="text-gray-700">Description</span>
+              <textarea
+                name="message"
+                className="
+            block
+            w-full
+            mt-2 px-16 py-8
+            border-gray-300
+            rounded-md
+            shadow-sm
+            focus:border-indigo-300
+            focus:ring
+            focus:ring-indigo-200
+            focus:ring-opacity-50
+          "
+                rows="5"
+                onChange={handleDescription}
+              ></textarea>
+            </label>
+          </div>
+
+          <div className="mb-2">
+            <label>
+              <span class="text-gray-700">Comment</span>
+              <textarea
+                name="message"
+                className="
+            block
+            w-full
+            mt-2 px-16 py-8
+            border-gray-300
+            rounded-md
+            shadow-sm
+            focus:border-indigo-300
+            focus:ring
+            focus:ring-indigo-200
+            focus:ring-opacity-50
+          "
+                rows="5"
+                onChange={handleComment}
+              ></textarea>
+            </label>
+          </div>
+
+          <div class="mb-6">
+            <button
+              type="submit"
+              className="
+            h-10
+            px-5
+            text-indigo-100
+            bg-indigo-700
+            rounded-lg
+            transition-colors
+            duration-150
+            focus:shadow-outline
+            hover:bg-indigo-800
+          "
+          onClick={handleSave}
+            >
+              Save
+            </button>
+          </div>
+          <div class="mb-6">
+            <button
+              type="submit"
+              className="
+            h-10
+            px-5
+            text-indigo-100
+            bg-indigo-700
+            rounded-lg
+            transition-colors
+            duration-150
+            focus:shadow-outline
+            hover:bg-indigo-800
+          "
+          onClick={() => closeModal(false)}
+            >
+              Cancel
+            </button>
+          </div>
+
+          <div></div>
+        </form>
+
+        <div className="bg-gray-50 px-4 py-3 sm:px-6 sm:flex sm:flex-row-reverse">
         </div>
+      </div>
+    </div>
+  </div>
     )
-}
 
-export default CreateIssueModal
+};
 
+export default CreateIssueModal;
